src: move image generation into separate file
This helps better organize things and makes the code a bit easier to
work with/focus on.
