+++ src/main.rs
@@ -12,19 +12,17 @@ extern crate urlencoded;
 use axum::{
     body::Bytes,
     error_handling::HandleErrorLayer,
-    extract::{Extension, Path, Query},
-    http::{header, HeaderMap, HeaderValue, StatusCode},
-    response::{ErrorResponse, Html, IntoResponse, Response},
+    extract::{Path, Query},
+    http::{header, HeaderValue, StatusCode},
+    response::{Html, IntoResponse},
     routing::get,
     BoxError, Router,
 };
 use handlebars::Handlebars;
-use image::{DynamicImage, ImageBuffer, ImageOutputFormat};
-use num::complex::Complex;
+use image::ImageBuffer;
 use redis::Commands;
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
-use std::io;
 use std::{
     collections::HashMap,
     net::SocketAddr,
@@ -38,6 +36,8 @@ use tower_http::{
     LatencyUnit, ServiceBuilderExt,
 };
 use urlencoded::UrlEncodedQuery;
+
+mod img;
 
 fn fetch_an_integer(key: &str, inc: bool) -> redis::RedisResult<i64> {
     // connect to redis
@@ -72,34 +72,6 @@ fn fetch_an_integer(key: &str, inc: bool) -> redis::RedisResult<i64> {
 //         Ok(res)
 //     }
 // }
-
-// ImgWriter writes a generated image out to the request
-struct ImgWriter {
-    img: ImageBuffer<image::Luma<u8>, Vec<u8>>,
-}
-
-// impl Modifier<Response> for ImgWriter {
-//     fn modify(self, res: &mut Response) {
-//         res.body = Some(Box::new(self));
-//     }
-// }
-
-impl IntoResponse for ImgWriter {
-    // fn write_body(&mut self, res: &mut dyn io::Write) -> io::Result<()> {
-    fn into_response(self) -> Response {
-        // Write to intermediary buffer because Seek is required.
-        let mut bytes: Vec<u8> = Vec::new();
-        DynamicImage::ImageLuma8(self.img.clone())
-            .write_to(&mut io::Cursor::new(&mut bytes), ImageOutputFormat::Png)
-            .map_err(|err| io::Error::new(io::ErrorKind::Other, err))
-            .unwrap();
-
-        let mut headers = HeaderMap::new();
-        headers.insert(header::CONTENT_TYPE, HeaderValue::from_static("image/png"));
-
-        (StatusCode::OK, headers, bytes).into_response()
-    }
-}
 
 static BASE: &'static str = r#"
 <!DOCTYPE html>
@@ -198,7 +170,7 @@ async fn main() {
     async fn peek_handler(
         Path(id): Path<String>,
         params: Query<HashMap<String, String>>,
-    ) -> Result<ImgWriter, StatusCode> {
+    ) -> Result<img::ImgWriter, StatusCode> {
         {
             let _ = match fetch_an_integer(id.as_str(), true) {
                 Ok(v) => v,
@@ -213,18 +185,18 @@ async fn main() {
         // let content_type = "image/png".parse::<Mime>().unwrap();
 
         match params.get("t").map(|t| t.as_str()) {
-            Some("mandelbrot") => apply_mandelbrot(&mut img, 500),
-            Some("julia") => apply_julia(&mut img, 500),
+            Some("mandelbrot") => img::apply_mandelbrot(&mut img, 500),
+            Some("julia") => img::apply_julia(&mut img, 500),
             Some(_) => return Err(StatusCode::NOT_FOUND),
             None => {
                 // turn the image white
                 let mut img_sm = ImageBuffer::new(1, 1);
-                apply_color(&mut img_sm, 255);
-                return Ok(ImgWriter { img: img_sm });
+                img::apply_color(&mut img_sm, 255);
+                return Ok(img::ImgWriter { img: img_sm });
             }
         };
 
-        Ok(ImgWriter { img: img })
+        Ok(img::ImgWriter { img: img })
     }
 
     async fn peek_info_handler(Path(id): Path<String>) -> Result<Html<String>, StatusCode> {
@@ -263,11 +235,11 @@ async fn main() {
         // Add high level tracing/logging to all requests
         .layer(
             TraceLayer::new_for_http()
-                // .on_body_chunk(|chunk: &Bytes, latency: Duration, _: &tracing::Span| {
-                //     tracing::trace!(size_bytes = chunk.len(), latency = ?latency, "sending body chunk")
-                // })
-                // .make_span_with(DefaultMakeSpan::new().include_headers(true))
-                // .on_response(DefaultOnResponse::new().include_headers(true).latency_unit(LatencyUnit::Micros)),
+                .on_body_chunk(|chunk: &Bytes, latency: Duration, _: &tracing::Span| {
+                    tracing::trace!(size_bytes = chunk.len(), latency = ?latency, "sending body chunk")
+                })
+                .make_span_with(DefaultMakeSpan::new().include_headers(true))
+                .on_response(DefaultOnResponse::new().include_headers(true).latency_unit(LatencyUnit::Micros)),
         )
         // Handle errors
         .layer(HandleErrorLayer::new(handle_errors))
@@ -301,71 +273,4 @@ async fn main() {
         .serve(app.into_make_service())
         .await
         .unwrap();
-}
-
-fn apply_color(img: &mut ImageBuffer<image::Luma<u8>, Vec<u8>>, value: u8) {
-    for (_, _, pixel) in img.enumerate_pixels_mut() {
-        *pixel = image::Luma([value]);
-    }
-}
-
-fn apply_mandelbrot(img: &mut ImageBuffer<image::Luma<u8>, Vec<u8>>, max_iters: i64) {
-    // from: https://en.wikipedia.org/wiki/Mandelbrot_set#Escape_time_algorithm
-
-    let scalex = 3.5 / img.width() as f32;
-    let scaley = 2.0 / img.height() as f32;
-
-    for (x, y, pixel) in img.enumerate_pixels_mut() {
-        let x0 = x as f32 * scalex - 2.5;
-        let y0 = y as f32 * scaley - 1.0;
-
-        let mut z = Complex::new(x0, y0);
-        // let c = z.clone();
-
-        let mut iters = 0;
-        for _ in 0..max_iters {
-            if z.norm() > 2.0 {
-                break;
-            }
-            let xt = z.re * z.re - z.im * z.im + x0;
-            z.im = 2.0 * z.re * z.im + y0;
-            z.re = xt;
-
-            // for some reason this is about 2x slower
-            // z = z * z + c;
-
-            iters += 1;
-        }
-
-        *pixel = image::Luma([iters as u8]);
-    }
-}
-
-fn apply_julia(img: &mut ImageBuffer<image::Luma<u8>, Vec<u8>>, max_iters: i64) {
-    // from: https://github.com/PistonDevelopers/image#62-generating-fractals
-
-    let scalex = 4.0 / img.width() as f32;
-    let scaley = 4.0 / img.height() as f32;
-
-    for (x, y, pixel) in img.enumerate_pixels_mut() {
-        let cy = y as f32 * scaley - 2.0;
-        let cx = x as f32 * scalex - 2.0;
-
-        let mut z = Complex::new(cx, cy);
-        let c = Complex::new(-0.4, 0.6);
-
-        let mut i = 0;
-
-        for t in 0..max_iters {
-            if z.norm() > 2.0 {
-                break;
-            }
-            z = z * z + c;
-            i = t;
-        }
-
-        // Create an 8bit pixel of type Luma and value i
-        // and assign in to the pixel at position (x, y)
-        *pixel = image::Luma([i as u8]);
-    }
 }
